Use base handler for sync folder job
